refactor: security 패키지 명 generator로 변경 및 api 주소 하이픈을 통해 표현 및 자잘한 리뷰 반영
